modularize stage 3
