+++ model/src/main/java/com/expediagroup/streamplatform/streamregistry/model/Producer.java
@@ -1,0 +1,36 @@
+/**
+ * Copyright (C) 2018-2019 Expedia, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.expediagroup.streamplatform.streamregistry.model;
+
+import java.util.List;
+
+import lombok.Builder;
+import lombok.NonNull;
+import lombok.Value;
+
+@Value
+@Builder
+public class Producer {
+  /**
+   * Name of the producer or Consumer
+   */
+  @NonNull String name;
+
+  /**
+   * Details of the producer or Consumer - region, cluster name, kafka configuration
+   */
+  @NonNull List<RegionStreamConfig> regionStreamConfigList;
+}
